test: consolidate crash-only JSONL coverage
